+++ packages/zent/__tests__/form/CreateForm-Field.js
@@ -69,9 +69,9 @@ describe('CreateForm and Field', () => {
     expect(typeof wrapper.context('zentForm')).toBe('object');
     expect(wrapper.state('_value')).toBe('');
     expect(wrapper.state('_isValid')).toBe(true);
-    expect(wrapper.state('_isPristine')).toBe(true);
+    expect(wrapper.state('_isDirty')).toBe(false);
     expect(wrapper.state('_isValidating')).toBe(false);
-    expect(wrapper.state('_pristineValue')).toBe('');
+    expect(wrapper.state('_initialValue')).toBe('');
     expect(wrapper.state('_validationError').length).toBe(0);
     expect(wrapper.state('_externalError')).toBe(null);
   });
@@ -134,8 +134,7 @@ describe('CreateForm and Field', () => {
     expect(
       Object.keys(wrapper.find('component').prop('validationErrors')).length
     ).toBe(0);
-    expect(wrapper.find('component').prop('isTouched')).toBe(false);
-    expect(wrapper.find('component').prop('isPristine')).toBe(true);
+    expect(wrapper.find('component').prop('isDirty')).toBe(false);
     expect(wrapper.find('component').prop('isValid')).toBe(true);
     expect(wrapper.find('component').prop('value')).toBe('');
     expect('value' in wrapper.find('component').props()).toBe(true);
@@ -306,6 +305,88 @@ describe('CreateForm and Field', () => {
       foo: 1
     });
     expect(wrapper.find(Field).getNode().state._value).toBe(1);
+    expect(wrapper.state('isFormValid')).toBe(true);
+  });
+
+  it('CreatedForm has initialize method which will be excuted with another revalidate', () => {
+    class FormForTest extends React.Component {
+      render() {
+        return (
+          <Form>
+            <Field
+              name="foo"
+              component={() => <div className="foo-div" />}
+              validations={{ required: true }}
+              value={1}
+            />
+          </Form>
+        );
+      }
+    }
+
+    const CreatedForm = createForm()(FormForTest);
+    const wrapper = mount(<CreatedForm />);
+    expect(wrapper.state('isFormValid')).toBe(true);
+    expect(wrapper.find(Field).getNode().state._value).toBe(1);
+    expect(wrapper.find(Field).getNode().state._initialValue).toBe(1);
+    wrapper.getNode().initialize({
+      foo: 12
+    });
+    expect(wrapper.find(Field).getNode().state._value).toBe(12);
+    expect(wrapper.find(Field).getNode().state._initialValue).toBe(12);
+    expect(wrapper.state('isFormValid')).toBe(true);
+    wrapper.getNode().reset({
+      foo: ''
+    });
+    expect(wrapper.find(Field).getNode().state._value).toBe('');
+    expect(wrapper.find(Field).getNode().state._initialValue).toBe(12);
+    expect(wrapper.state('isFormValid')).toBe(false);
+    wrapper.getNode().initialize();
+    expect(wrapper.find(Field).getNode().state._value).toBe(12);
+    expect(wrapper.find(Field).getNode().state._initialValue).toBe(12);
+    expect(wrapper.state('isFormValid')).toBe(true);
+  });
+
+  it('CreatedForm has setFieldsValue method which will be excuted with another revalidate', () => {
+    class FormForTest extends React.Component {
+      render() {
+        return (
+          <Form>
+            <Field
+              name="foo"
+              component={() => <div className="foo-div" />}
+              validations={{ required: true }}
+              value={1}
+            />
+            <Field
+              name="bar"
+              component={() => <div className="bar-div" />}
+              validations={{ required: true }}
+              value={2}
+            />
+          </Form>
+        );
+      }
+    }
+
+    const CreatedForm = createForm()(FormForTest);
+    const wrapper = mount(<CreatedForm />);
+    expect(wrapper.state('isFormValid')).toBe(true);
+    expect(wrapper.getNode().fields[0].state._value).toBe(1);
+    expect(wrapper.getNode().fields[0].state._initialValue).toBe(1);
+    expect(wrapper.getNode().fields[0].state._isDirty).toBe(false);
+    expect(wrapper.getNode().fields[1].state._value).toBe(2);
+    expect(wrapper.getNode().fields[1].state._initialValue).toBe(2);
+    expect(wrapper.getNode().fields[1].state._isDirty).toBe(false);
+    wrapper.getNode().setFieldsValue({
+      foo: 12
+    });
+    expect(wrapper.getNode().fields[0].state._value).toBe(12);
+    expect(wrapper.getNode().fields[0].state._isDirty).toBe(true);
+    expect(wrapper.getNode().fields[0].state._initialValue).toBe(1);
+    expect(wrapper.getNode().fields[1].state._value).toBe(2);
+    expect(wrapper.getNode().fields[1].state._initialValue).toBe(2);
+    expect(wrapper.getNode().fields[1].state._isDirty).toBe(false);
     expect(wrapper.state('isFormValid')).toBe(true);
   });
 
@@ -540,10 +621,10 @@ describe('CreateForm and Field', () => {
     wrapper.unmount();
     wrapper.mount();
     expect(wrapper.state('isFormValid')).toBe(false);
-    const external = wrapper.getNode().setFieldExternalErrors;
-    expect(() => {
-      external({ foo: 'bar', bar: 321 });
-    }).toThrow();
+    // const external = wrapper.getNode().setFieldExternalErrors;
+    // expect(() => {
+    //   external({ foo: 'bar', bar: 321 });
+    // }).toThrow();
     wrapper = mount(<CreatedForm />);
     wrapper.setProps({ hackSwitch: true });
     wrapper.getNode().setFieldExternalErrors({ foo: 'bar', bar: 321 });