+++ site/src/nav.config.js
@@ -3,9 +3,10 @@
 module.exports = {
   'zh-CN': [
     {
-      name: '使用指南',
+      name: '指南',
       groups: [
         {
+          groupName: '使用',
           list: [
             {
               title: '快速上手',
@@ -13,10 +14,20 @@ module.exports = {
               component: require('../../packages/zent/README.md')
             },
             {
+              title: '项目示例',
+              path: ' guides/demos',
+              component: require('../../packages/zent/docs/demo.md')
+            },
+            {
               title: 'babel-plugin-zent',
               path: 'guides/babel-plugin-zent',
               component: require('../../packages/babel-plugin-zent/README.md')
-            },
+            }
+          ]
+        },
+        {
+          groupName: '主题',
+          list: [
             {
               title: '色彩',
               path: 'guides/colors',
@@ -26,7 +37,12 @@ module.exports = {
               title: '定制主题',
               path: 'guides/theme',
               component: require('../../packages/zent/docs/theme.md')
-            },
+            }
+          ]
+        },
+        {
+          groupName: '升级',
+          list: [
             {
               title: '更新日志',
               path: 'guides/changelog',
@@ -38,7 +54,22 @@ module.exports = {
               component: require('../../packages/zent/CHANGELOG.md')
             },
             {
-              title: '贡献代码',
+              title: '3.x 升级指南',
+              path: 'migrating/3x',
+              component: require('../../packages/zent/docs/migrating-2-3.md')
+            },
+            {
+              title: '2.1.x 升级指南',
+              path: 'migrating/21x',
+              component: require('../../packages/zent/docs/migrating-2.1.md')
+            }
+          ]
+        },
+        {
+          groupName: '开发',
+          list: [
+            {
+              title: '如何参与',
               path: 'guides/contribute',
               component: require('../../packages/zent/docs/CONTRIBUTING.md')
             },
@@ -46,26 +77,6 @@ module.exports = {
               title: '文档规范',
               path: 'guides/markdown',
               component: require('../../packages/zent/docs/MARKDOWN.md')
-            }
-          ]
-        }
-      ]
-    },
-
-    {
-      name: '升级指南',
-      groups: [
-        {
-          list: [
-            {
-              title: '2.1.x 升级指南',
-              path: 'migrating/21x',
-              component: require('../../packages/zent/docs/migrating-2.1.md')
-            },
-            {
-              title: '3.x 升级指南',
-              path: 'migrating/3x',
-              component: require('../../packages/zent/docs/migrating-2-3.md')
             }
           ]
         }
@@ -104,6 +115,11 @@ module.exports = {
               component: require('../../packages/zent/src/badge/README.md')
             },
             {
+              path: 'component/block-header',
+              title: 'BlockHeader 标题',
+              component: require('../../packages/zent/src/block-header/README.md')
+            },
+            {
               path: 'component/card',
               title: 'Card 卡片',
               component: require('../../packages/zent/src/card/README.md')
@@ -129,13 +145,18 @@ module.exports = {
               component: require('../../packages/zent/src/pop/README.md')
             },
             {
+              path: 'component/preview-image',
+              title: 'previewImage 图片预览',
+              component: require('../../packages/zent/src/preview-image/README.md')
+            },
+            {
               path: 'component/progress',
               title: 'Progress 进度条',
               component: require('../../packages/zent/src/progress/README.md')
             },
             {
               path: 'component/sweet-alert',
-              title: 'SweetAlert',
+              title: 'Sweetalert 快捷对话框',
               component: require('../../packages/zent/src/sweetalert/README.md')
             },
             {
@@ -147,11 +168,6 @@ module.exports = {
               path: 'component/tag',
               title: 'Tag 标签',
               component: require('../../packages/zent/src/tag/README.md')
-            },
-            {
-              path: 'component/preview-image',
-              title: 'previewImage 图片预览',
-              component: require('../../packages/zent/src/preview-image/README.md')
             }
           ]
         },
@@ -189,6 +205,11 @@ module.exports = {
               component: require('../../packages/zent/src/datetimepicker/README.md')
             },
             {
+              path: 'component/date-range-quick-picker',
+              title: 'DateRangeQuickPicker',
+              component: require('../../packages/zent/src/date-range-quick-picker/README.md')
+            },
+            {
               path: 'component/design',
               title: 'Design 微页面编辑',
               component: require('../../packages/zent/src/design/README.md')
@@ -220,7 +241,7 @@ module.exports = {
             },
             {
               path: 'component/select',
-              title: 'Select 选择器',
+              title: 'Select 下拉选择',
               component: require('../../packages/zent/src/select/README.md')
             },
             {
@@ -259,14 +280,19 @@ module.exports = {
               component: require('../../packages/zent/src/breadcrumb/README.md')
             },
             {
-              path: 'component/menu',
-              title: 'Menu 菜单',
-              component: require('../../packages/zent/src/menu/README.md')
-            },
-            {
               path: 'component/dropdown',
               title: 'Dropdown 下拉菜单',
               component: require('../../packages/zent/src/popover/Dropdown.md')
+            },
+            {
+              path: 'component/grid',
+              title: 'Grid 网格',
+              component: require('../../packages/zent/src/grid/README.md')
+            },
+            {
+              path: 'component/menu',
+              title: 'Menu 菜单',
+              component: require('../../packages/zent/src/menu/README.md')
             },
             {
               path: 'component/pagination',
@@ -282,11 +308,6 @@ module.exports = {
               path: 'component/table',
               title: 'Table 表格',
               component: require('../../packages/zent/src/table/README.md')
-            },
-            {
-              path: 'component/grid',
-              title: 'Grid 网格',
-              component: require('../../packages/zent/src/grid/README.md')
             },
             {
               path: 'component/tabs',
@@ -312,21 +333,11 @@ module.exports = {
               path: 'component/portal',
               title: 'Portal 传送门',
               component: require('../../packages/zent/src/portal/README.md')
-            }
-          ]
-        },
-        {
-          groupName: '业务组件',
-          list: [
-            {
-              path: 'component/block-header',
-              title: 'BlockHeader 标题',
-              component: require('../../packages/zent/src/block-header/README.md')
             },
             {
-              path: 'component/date-range-quick-picker',
-              title: 'DateRangeQuickPicker',
-              component: require('../../packages/zent/src/date-range-quick-picker/README.md')
+              path: 'component/InfiniteScroller',
+              title: 'InfiniteScroller 无限滚动',
+              component: require('../../packages/zent/src/infinite-scroller/README.md')
             }
           ]
         }